+++ cmd/ui/BUILD.bazel
@@ -5,7 +5,6 @@ go_library(
     srcs = [
         "contentView.go",
         "fileList.go",
-        "prStatusView.go",
         "prViewer.go",
         "pullRequestHeader.go",
         "statusBar.go",
@@ -21,7 +20,6 @@ go_library(
         "@com_github_charmbracelet_glamour//:glamour",
         "@com_github_charmbracelet_lipgloss//:lipgloss",
         "@com_github_erikgeiser_promptkit//confirmation",
-        "@com_github_evertras_bubble_table//table",
         "@com_github_pterm_pterm//:pterm",
         "@com_github_treilik_bubbleboxer//:bubbleboxer",
     ],
